Add adminSlice
