+++ src/admin/components/CreateEventForm/index.tsx
@@ -1,11 +1,11 @@
-import React, { ChangeEventHandler, FocusEventHandler, FormEventHandler } from 'react';
-import { Form, Input, Button, Select, DatePicker, TimePicker, Upload } from 'antd';
+import { Button, DatePicker, Form, Input, Select, TimePicker, Upload } from 'antd';
 import ImgCrop from 'antd-img-crop';
 import 'antd/es/modal/style';
 import 'antd/es/slider/style';
 import * as moment from 'moment';
+import React, { ChangeEventHandler, FocusEventHandler, FormEventHandler } from 'react';
 import { useHistory } from 'react-router-dom';
-
+import { copyLink } from '../../utils';
 import './style.less';
 
 const { Option } = Select;
@@ -45,12 +45,11 @@ interface CreateEventFormProps {
     attendanceCode: string | null;
     description: string | null;
   };
-  copyLink: Function;
 }
 
 /* Future Note: Add a fun generate attendance code function :) based on title */
 const CreateEventForm: React.FC<CreateEventFormProps> = (props) => {
-  const { handleBlur, handleChange, handleSubmit, setFieldTouched, setFieldValue, values, errors, copyLink } = props;
+  const { handleBlur, handleChange, handleSubmit, setFieldTouched, setFieldValue, values, errors } = props;
   const history = useHistory();
 
   return (
@@ -162,13 +161,7 @@ const CreateEventForm: React.FC<CreateEventFormProps> = (props) => {
           >
             Discard
           </Button>
-          <Button
-            type="primary"
-            className="link-button"
-            onClick={() => {
-              copyLink(values.attendanceCode);
-            }}
-          >
+          <Button type="primary" className="link-button" onClick={() => copyLink(values.attendanceCode)}>
             Copy Checkin Link
           </Button>
         </form>
